requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.0.4
dataclasses-json==0.5.14
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.4.0
gunicorn==21.2.0
heroku==0.1.4
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
langchain==0.0.248
langsmith==0.0.16
MarkupSafe==2.1.1
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.12
python-dateutil==1.5
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
yarl==1.9.2
